Move linter rule to separate PR. Move common functions to utils.js
