+++ src/core/figures.js
@@ -7,6 +7,7 @@
 import {
   addId,
   getIntlData,
+  getPreviousSections,
   renameElement,
   showWarning,
   wrapInner,
@@ -147,30 +148,5 @@ function decorateTableOfFigures(tofElement) {
     tofElement.classList.add("introductory");
   } else if (previousSections.some(sec => sec.classList.contains("appendix"))) {
     tofElement.classList.add("appendix");
-  }
-}
-
-/**
- * @param {Element} element
- */
-function getPreviousSections(element) {
-  /** @type {Element[]} */
-  const sections = [];
-  for (const previous of iteratePreviousElements(element)) {
-    if (previous.localName === "section") {
-      sections.push(previous);
-    }
-  }
-  return sections;
-}
-
-/**
- * @param {Element} element
- */
-function* iteratePreviousElements(element) {
-  let previous = element;
-  while (previous.previousElementSibling) {
-    previous = previous.previousElementSibling;
-    yield previous;
   }
 }
